Split array::mruby and array::trampoline into mruby, trampoline, and ffi modules
ffi module funcs cannot allocate on the Ruby heap.
