+++ artichoke-backend/src/extn/core/array/mod.rs
@@ -14,6 +14,7 @@ use crate::Artichoke;
 
 pub mod args;
 pub mod backend;
+pub mod ffi;
 pub mod mruby;
 pub mod trampoline;
